separate libs from core modules
